Convert calculated float to int

+++ ContadorObjetosEmMovimento.py
@@ -77,8 +77,8 @@ while True:
     QtdeContornos = 0
 
     #desenha linhas de referencia 
-    CoordenadaYLinhaEntrada = (height / 2)-OffsetLinhasRef
-    CoordenadaYLinhaSaida = (height / 2)+OffsetLinhasRef
+    CoordenadaYLinhaEntrada = int((height / 2)-OffsetLinhasRef)
+    CoordenadaYLinhaSaida = int((height / 2)+OffsetLinhasRef)
     cv2.line(Frame, (0,CoordenadaYLinhaEntrada), (width,CoordenadaYLinhaEntrada), (255, 0, 0), 2)
     cv2.line(Frame, (0,CoordenadaYLinhaSaida), (width,CoordenadaYLinhaSaida), (0, 0, 255), 2)
